makes it possible to run an effect in the first callback of 'CollectResponse'
A typical client would here want to be able to run extra computation in order to carry on and do whatever it needs to do to collect or constuct the next request to send. Being restricted to only pure computation here makes it fairly hard to pipelined another request unless the request can be constructed out of thin air or, has already been collected and buffered upfront.